fixup: move getFetcher into its own package
