+++ packages/node.fetcher/tsconfig.json
@@ -25,7 +25,7 @@
       "path": "../fetching-types"
     },
     {
-      "path": "../package-requester"
+      "path": "../pick-fetcher"
     },
     {
       "path": "../tarball-fetcher"
