+++ packages/package-requester/src/index.ts
@@ -2,7 +2,5 @@ import packageRequester from './packageRequester'
 
 export default packageRequester
 
-export { getFetcher } from './packageRequester'
-
 export { PackageFilesResponse } from '@pnpm/fetcher-base'
 export { PackageResponse } from '@pnpm/store-controller-types'
