+++ packages/package-requester/src/packageRequester.ts
@@ -10,6 +10,7 @@ import {
   PackageFilesIndex,
 } from '@pnpm/cafs'
 import { fetchingProgressLogger, progressLogger } from '@pnpm/core-loggers'
+import { getFetcher } from '@pnpm/pick-fetcher'
 import PnpmError from '@pnpm/error'
 import {
   Cafs,
@@ -672,34 +673,4 @@ async function fetcher (
     })
     throw err
   }
-}
-
-export function getFetcher (fetcherByHostingType: {[hostingType: string]: FetchFunction}, resolution: Resolution) {
-  let fetcherType = resolution.type
-
-  if (resolution.type == null) {
-    if (resolution.tarball.startsWith('file:')) {
-      fetcherType = 'localTarball'
-    } else if (isGitHostedPkgUrl(resolution.tarball)) {
-      fetcherType = 'gitHostedTarball'
-    } else {
-      fetcherType = 'remoteTarball'
-    }
-  }
-
-  const fetch = fetcherByHostingType[fetcherType!]
-
-  if (!fetch) {
-    throw new Error(`Fetching for dependency type "${resolution.type ?? 'undefined'}" is not supported`)
-  }
-
-  return fetch
-}
-
-function isGitHostedPkgUrl (url: string) {
-  return (
-    url.startsWith('https://codeload.github.com/') ||
-    url.startsWith('https://bitbucket.org/') ||
-    url.startsWith('https://gitlab.com/')
-  ) && url.includes('tar.gz')
 }
