+++ packages/pick-fetcher/README.md
@@ -1,0 +1,15 @@
+# @pnpm/pick-fetcher
+
+> Pick a tarball fetcher by type
+
+[![npm version](https://img.shields.io/npm/v/@pnpm/pick-fetcher.svg)](https://www.npmjs.com/package/@pnpm/pick-fetcher)
+
+## Deploy
+
+```sh
+pnpm add @pnpm/pick-fetcher
+```
+
+## License
+
+MIT
