+++ packages/pick-fetcher/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@pnpm/pick-fetcher",
+  "version": "0.0.0",
+  "description": "Pick a tarball fetcher by type",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "files": [
+    "lib",
+    "!*.map"
+  ],
+  "scripts": {
+    "lint": "eslint src/**/*.ts test/**/*.ts",
+    "_test": "jest",
+    "test": "pnpm run compile && pnpm run _test",
+    "prepublishOnly": "pnpm run compile",
+    "compile": "tsc --build && pnpm run lint --fix"
+  },
+  "repository": "https://github.com/pnpm/pnpm/blob/main/packages/pick-fetcher",
+  "license": "MIT",
+  "engines": {
+    "node": ">=14.6"
+  },
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "homepage": "https://github.com/pnpm/pnpm/blob/main/packages/pick-fetcher#readme",
+  "funding": "https://opencollective.com/pnpm",
+  "keywords": [
+    "pnpm7"
+  ],
+  "devDependencies": {
+    "@pnpm/fetcher-base": "workspace:*",
+    "@pnpm/pick-fetcher": "workspace:*",
+    "@pnpm/resolver-base": "workspace:*"
+  },
+  "exports": {
+    ".": "./lib/index.js"
+  }
+}
