+++ packages/pick-fetcher/tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "extends": "@pnpm/tsconfig",
+  "compilerOptions": {
+    "outDir": "lib",
+    "rootDir": "src"
+  },
+  "include": [
+    "src/**/*.ts",
+    "../../typings/**/*.d.ts"
+  ],
+  "references": [
+    {
+      "path": "../fetcher-base"
+    },
+    {
+      "path": "../resolver-base"
+    }
+  ]
+}
